data-and-ai-jhb/oct/code/data-ai/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
azure-ai-projects==1.0.0b10
azure-identity==1.19.0
jinja2==3.1.4
python-multipart==0.0.20
